+++ tests/lib/rules/jsx-key.js
@@ -176,6 +176,33 @@ ruleTester.run('jsx-key', rule, {
       `,
       features: ['types', 'no-babel-old'],
     },
+    { code: 'React.Children.toArray([1, 2 ,3].map(x => <App />));' },
+    {
+      code: `
+        import { Children } from "react";
+        Children.toArray([1, 2 ,3].map(x => <App />));
+      `,
+    },
+    {
+      // TODO: uncomment the commented lines below
+      code: `
+        import Act from 'react';
+        import { Children as ReactChildren } from 'react';
+        
+        const { Children } = Act;
+        const { toArray } = Children;
+        
+        Act.Children.toArray([1, 2 ,3].map(x => <App />));
+        Act.Children.toArray(Array.from([1, 2 ,3], x => <App />));
+        Children.toArray([1, 2 ,3].map(x => <App />));
+        Children.toArray(Array.from([1, 2 ,3], x => <App />));
+        // ReactChildren.toArray([1, 2 ,3].map(x => <App />));
+        // ReactChildren.toArray(Array.from([1, 2 ,3], x => <App />));
+        // toArray([1, 2 ,3].map(x => <App />));
+        // toArray(Array.from([1, 2 ,3], x => <App />));
+      `,
+      settings,
+    },
   ]),
   invalid: parsers.all([
     {